Implement call at bytecode level
We now have a way to store a pointer to a function in a JS variable
and call it. The function can accept any number of arguments.

Currently two types of functions can be called:
- functions implemented in froth: e.g `jsprint` (bound to the `print`
  global).
- low level functions implemented in C: they have access to the
  interpreter and have to deal with the stack manually.

The next step is to implement a lightweight FFI interface, which will
be expose the same froth ABI, but cause an arbitrary C function to be
called and have the FFI layer deal with taking the arguments from the
froth stack, converting them to the appropriate type and making the
actual call.

PUBLISHED_FROM=71ca8ed5324319b1818f470f0468fa686c79081a