Add more comments to jscall thing
And refactor it slightly: now, `jscall` is a wrapper around another
call, because when the function to call is a BF function, then
everything after calling the "old" jscall is ignored, because of the way
jsexit restores context. That may be confusing, so now `jscall` behaves
in a conventional way.

PUBLISHED_FROM=77fc147cfd09ab9ef5a809d46e996be18022081f