Fix ffi bool support
Compilers can generate different code for functions returning `int` and
those returning `bool`. E.g. for x86, this is the disassembly of
gcc-generated non-optimized code for the function returning `int`:

```
588     int ffi_test_bi(int a0) {
   0x0000000000449e02 <+0>:     push   %rbp
   0x0000000000449e03 <+1>:     mov    %rsp,%rbp
   0x0000000000449e06 <+4>:     mov    %edi,-0x4(%rbp)

589       return a0 > 50;
   0x0000000000449e09 <+7>:     cmpl   $0x32,-0x4(%rbp)
   0x0000000000449e0d <+11>:    setg   %al
   0x0000000000449e10 <+14>:    movzbl %al,%eax

590     }
   0x0000000000449e13 <+17>:    pop    %rbp
   0x0000000000449e14 <+18>:    retq
```

And this is for function returning `bool`:
```
588     bool ffi_test_bi(int a0) {
   0x0000000000449e02 <+0>:     push   %rbp
   0x0000000000449e03 <+1>:     mov    %rsp,%rbp
   0x0000000000449e06 <+4>:     mov    %edi,-0x4(%rbp)

589       return a0 > 50;
   0x0000000000449e09 <+7>:     cmpl   $0x32,-0x4(%rbp)
   0x0000000000449e0d <+11>:    setg   %al

590     }
   0x0000000000449e10 <+14>:    pop    %rbp
   0x0000000000449e11 <+15>:    retq
```

So in case of bool, the value is left in the register `al` only. Clang
behaves in the same way, but only if optimization is on.

So, this commit creates a separate set of functions which return bool.

PUBLISHED_FROM=b3654f291b1c362f2c8345eb0e22aad79ed65e98