Fix mjs_set and mjs_set_v
The biggest problem (a bug) was that `mjs_set_v` was working by
obtaining a string pointer from the provided `mjs_val_t` (and the
pointer was pointing to some place within the mjs strings buffer), and
calling `mjs_set` with that pointer. Then, `mjs_set` was creating a new
string, and if it happens that the buffer needs to be resized and
relocated, the provided pointer was becoming invalid.

And a smaller problem is that calling `mjs_set_v` caused the provided
string (if provided `mjs_val_t` was indeed a string) to be duplicated.

So now both `mjs_set` and `mjs_set_v` call the internal worker function
which does not do unnecessary string duplication, and thus this
eliminates the bug.

PUBLISHED_FROM=d3a06fc8265ec01d0fa32469ae1db1ca0b4c5fba